Fix #106 - No more entering your password twice at wallet open
We cache the passwords that the 'InstallWizard' got frmo the user for 10
seconds.  This means in the UI case where the user opened a wallet that
is cashshuffle enabled in the settings, they will not have to enter
their password twice.

In other UI cases: such as enabling for the first time or
enable-then-disable-then-enable again, passwords are still required.