Cash Address: Allow hide of status button in preferences + refactor
The user may elect to disable/hide the cash address toggle button in
preferences.  See Preferences -> General.

In addition the Preferences -> General tab has been laid out a little
bit differently with the Address format set as a combo box rather than a
checkbox (this is more to allow for future/SLP).

A checkbox was also added to Preferences -> General to allow hiding of
the status bar button app-wide (in case the user wants to avoid status
bar clutter). The status bar button is shown on new installs/by default.

Additionally, internal code refactorings are in this commit to move the
'cashaddr_toggled_signal' from a per-window signal to an app-global
signal (moved to ElectrumGui instance).  It makes more sense for it to
be a single app-wide signal as it's really an app-wide setting.

An additional signal was added, cashaddr_status_button_hidden_signal as
an app-wide signal for when the status button gets hidden (all windows
listen for this signal and act accordingly).