CashAccounts: Fix: closing wallet before confirm -> no [mine] in contacts
The bug was because we were filtering out "incomplete" scripts on wallet
load. We shouldn't be doing this.  Was an oversight on the programmer's
(my) part.

The bug is now fixed. You can safely close a wallet after registration
and opening it later will show your verified registration in the contacts tab