CashAccounts: Implemented Cash Account -> Details...
Accessible from the Addresses tab if you right click on an address that
has a Cash Account, and you go into the Cash Account's own submenu, and
you select "Details..." on it.

Also accessible from the Contacts tab, right-click -> "Details..." on a
Cash Account contact.

The details screen just shows the emoji in a large font and the Cash
Account with a large font, and also a QR code for the Address and Cash
Account.

It doesn't add any real functionality, but is just a UI nicety to make
the Cash Account feel more "real" by seeing the big emoji and the big
bold title text.

This facility is modeled after the cashaccount.info results page
layout.