CashAccounts: Fixup for cashacctqt.cash_account_detail_dialog
Enforce the passed-in cash account is of a known type, and display an
error message if it is not.

This is preferable to assert (which is what we had before) which leads
to the crash reporter when a user error message will do.

Normally in current code this situation cannot be encountered but this
change future-proofs the function in case callers call it with cash
accounts that are of an unsupported type.