wallet & Qt: Fixed erroneous coinbase maturity logic
Coinbase maturity is calculated based off chainTip->nHeight + 1 (ie
mempoolHeight) for allowing a tx into the mempool and into the next
block.  Thus there need to be 99 intervening blocks between a coinbase
reward and when it can be mined (it can be mined on the 100th block from
when it was created, and it can make it into the mempool after 99 blocks
are mined on top of it).

The extant code got it wrong and was off by 1 in the positive direction
(that is, it waited an extra block).

From bitcoind txmempool.cpp:

 if (coin->IsSpent() ||
     (coin->IsCoinBase() && ((signed long)nMemPoolHeight) - coin->nHeight < COINBASE_MATURITY))
 {
     transactionsToRemove.push_back(*tx);
     break;
 }

From bitcoind's consensus/tx_verify.cpp:

if (nSpendHeight - nCoinHeight < COINBASE_MATURITY)
    return state.Invalid(false, REJECT_INVALID, "bad-txns-premature-spend-of-coinbase",
        strprintf("tried to spend coinbase at depth %d", nSpendHeight - nCoinHeight));

Above, nSpendHeight is basically chainTip()->nHeight + 1, and
nMemPoolHeight is the same.

Thus, if COINBASE_MATURITY were 1, you could spend a coinbase coin right
away.

At 100, you can spend it after 99 intervening blocks.

This commit fixes that.

Also in this commit: Hightlight coinbase immature coins in the UTXOList
(Coins) tab, color them blue, and indicate they can't be spent.