Qt: Fix keyPressEvent not calling super() when accepting
This in particular affected the util.py EnterButton class.  As such, the
button was eating keypresses. This commit fixes that. As per the Qt
docs, if you don't want to eat a keyPressEvent, you should call the
superclass keyPressEvent handler to pass it on.

Also in this commit: prefer super().keyPressEvent(e) over
MyBaseClass.keyPressEvent(self, e).