Build: Make build scripts usable for cross compilation
Instead of duplicating the build logic in the Windows build scripts, this
makes the make_* build scripts work for cross compilation by automatically
detecting the platform specific output file name. The scripts were made a
bit simpler to adjust to new packages as well.

We also skip messing with the submodules and add a make_clean script
instead. This makes it easier to work in the submodules without losing
work.

There is a new host_strip function that calls the appropriate program
to strip a host binary.