Server list: Add 'display' key to give Tor servers a short name
New .onion URLs are quite long and it is hard to remember which
server is which. This patch adds a 'display' key that can be added
to Tor servers in servers.json to give them a short name. Such a
server is then listed as "<display> (.onion)" and gets a Tor icon
in the server list.