Fixed the long-standing "tx's lose their txid" bug
It turns out in some cases (particularly when dong spend from -> to send
tab), the tx in question would point to mutable data that somehow, for
some resaon, the send tab would modify and make it lose sigs after
broadcast.

This fixes that bug, by taking a deep copy of the tx before broadcast.