Fix #1785 - Allow searching (Ctrl+F) by tx_hash in history list
To accomplish this we added another list to MyTreeWidget which
can be set by subclasses `filter_data_columns`.  By default this list is
empty (and thus data column searching is disabled).  If specified in
subclasses, in addition to the usual item.text() searches, data columns
may be searched as well.

Only the HistoryList class uses this facility currently to allow for
searching by tx_hash.