Wallet.py and keystore.py fixups
Since we recently added some "derivations" fields and other miscellany
to the keystore format, we need to ensure they get preserved if the user
switches EC versions to a previous version.  We added code to save the
derivations to top-level storage 'keystore_derivation'.  On wallet
reload with a later EC version, these derivations are restored and put
back in the keystore instance(s) for the wallet.

Also, we renamed the keystore field 'is_seed_bip39' to 'seed_type'.  The
current seed_types are: "old", "electrum", and "bip39".

Also a bugfix -- electron-cash CLI 'create' command stopped working due
to recent commits; fixed.