OSX: Release .app startup time improved significantly
We went with an flat app bundle containing all files, rather than a
onefile as we had previously.

This plays much nicer with macOS Catalina GateKeeper letting the app
launch instantly. The previous onefile we were creating was causing
the OS to phone home to make sure the _MEIxxx extracted bundle was
kosher.  So -- to avoid all the slowdown, we don't extract any bundles.
Instead, the program  executes directly from the .app.

So now Electron Cash on macOS Catalina is finally a first class citizen!