Tx signing: Follow-up, hook in use_cache mechanism to UI signing
Now the UI send tab as well as the tx dialog both make use of the
"use_cache" mechanism to significantly speed up tx signing.  In both
thesse places the tx is immutable once the user gets to the signing
phase, so it's safe to use the common sig cache for the tx in that case.

Also both UI places take a deep copy of the tx, so should the tx be
modified later (in the case of the send tab), this is safe.