ZBar: Fix Windows build DLL imports
When improving the build scripts to add Tor support we forgot to
transfer some Windows specific flags into the new ZBar build scripts,
which resulted in the binary depending on additional .dll files that
were not bundled into the installer, causing the QR scanner not to
work for most users. This patch restores the flags.