CashFusion: Fixed some issues caused by the `mpz` type from ecdsa
Fuses would fail if user had `gmpy2` installed with `ecdsa` > 0.14. In
that case ecdsa works with `mpz` as the type, rather than `int`, which
is not 100% compatible with int.

This works around that problem by casting to int each time it does e.g.
to_bytes().

Seems to work.  I may have missed a spot.. but such is the misery of
lack of static types.