Tests: Fix failing host name test and add wildcard tests
In 4.1.0 we added support for wildcard certificates which caused
the tests to fail. The test was passing data in a format
ssl.match_hostname did not expect, so the test had to be fixed.

This also adds tests for wildcard certificates and tests the second
level sub domains of wildcard certificates fail the test.