[Qt] Set the layout direction based on user-configured language
For Arabic and Farsi (and Hebrew and others), the Qt layout direction
was wrong.  Layout direction needs to be set to right-to-left for these
languages, resulting in basically a flipped window (which is the
standard for desktop apps using these languages).

This fix makes the Qt app render correctly in these languages.

Special thanks to Akad for calling this to my attention.