Qt UI: Made all tabs except for "Console" on by default
On new installs, we had "Addresses", "Coins" and "Contacts" disabled.
But since Electron Cash is a power wallet, we decided it's best for new
users to see all the tabs on first run, rather than having to manually
enable these useful tabs.

So they have been enabled by default now on new installs.

Since CashFusion is so popular, it is useful to be able to see one's
coins now and one's addresses from the get-go.