TxDialog: Add "Freeze Coins" button to freeze inputs before broadcast
This new button allows users to freeze inputs for a signed transaction
before broadcast.  This is most useful for users wishing to broadcast
a signed transaction later.  Freezing the inputs to a transaction
ensures that the coins aren't spent before broadcast.

This facility is intended for multisig wallet users but is not
necessarily only useful for them.

Also in this commit: Various nits and refactorings, added some tooltips
to the TxDialog, and other small code fixups.