fix deadlock in slp_mgt tab
In issue #144 a user reported very poor performace in this part
of the user interface.  This commit fixes at least part of the issues
experienced in the bug report.

I've refactored the code to avoid the deadlock and address
domain for inner call to get_slp_utxos.  The fact that the tokens
tab ever worked at all with the way it was written is impressive.