Fixed a regression introduced in 4.2.1
The issues manifests itself whenever the checkpoint block is after
the asert anchor. This caused the code in blockchain.py to have a bad time.

The asert anchor calculation code needs all headers, but if we have a
checkpoint after the anchor, all headers are not always available. The fix
is to either have checkpoints before the anchor, or to hard-code the anchor.

This is the case for all chains we support now.

We hard-coded the anchor for all chains except for scalenet, which will
reorg back to before the anchor each time so hard-coding the anchor
would break things in the future.