Performance tweak: Don't check is_mine when freezing coins
The check was to ensure that as little "junk" ends up in the frozen coin
set as possible.  However currently all code paths that lead to
`set_frozen_coin_state` always freeze coins in the wallet and never any
non-wallet coins.

Since the `is_mine` check is not free, and since some users may be
freezing lots of coins at once on large wallets, the `is_mine` check can
be removed. This turns an O(N log M) (where N is the number of coins and
M is the size of the address set) into a simple O(N) loop.

In my testing on a wallet with 40000 utxos and 1500 addresses,
freezing/unfreezing went from 90ms to 45 ms. Not a huge savings but not
zero either. Maybe it will make a difference to some plugin such as
CashFusion.

Note that it's not a huge tragedy to freeze a coin that is not `is_mine`
-- it will just take up a little extra space in the wallet file forever.
However since ending up in such a situation is currently impossible in
this codebase, it's better to not check `is_mine` on every iteration when
freezing coins.

Partially addresses #2102.