NSIS: Create installer log using AdvUninstLog.nsh
The installer has a problem with removing its files / directories
recursively. This risked users accidentally deleting their wallets if
they were stored in the installation directory. To protect against this
a check was added to ask the user to confirm the deletion. This in turn
broke silent installations because the prompt couldn't be skipped.

To fix this we add `AdvUninstLog.nsh`, which allows the installer to log
all the files it is installing and then only delete the files it
actually installed when uninstalling. We also add an uninstallation
prompt when an existing installation is detected.

The `AdvUninstLog.nsh` file was slightly changed to silence a warning.
In the `UNINSTALL.LOG_CLOSE_INSTALL` macro `!undef ID ${__LINE__}` was
replaced with `!undef ID`.

This also switches the installer to Unicode, ANSI NSIS builds are now
deprecated.