disable main_window.py experimental PO components
These will be updated in subsequent comments, as the PO is currently
experimental.

If we do the main_window components like this, the checkbox needs
a signal to syncronize with the Network dialog checkbox.  I'm also
wondering if this is what we want to display on the main window,
maybe it can be simplified..

For example, if the user has no BCH and goes to send tokens, then a
modal dialog can pop up to ask the user if he wants to enable the post
office.  And whenever the post office is going to be used, as determined
by the calculate fee loop, a special icon can be shown to indicate whether
or not the post office is going to be used in the send.

We should also allow the user to click "Max" with the postage
automatically subtracted from the max amount with an indicator
showing how much the postage fee is.