fix nft child validation restart
This is to follow-up to the previous commit that enabled graph search
for nft groups (2d2e3dd).  After that commit, there was some cases
where NFT child validation would get stuck because the paused child
job wasn't added to the list of paused job (bc it was the current job).
So to handle this we call start() on the job directly if it's not in the
list of paused validation jobs.

Also re-enabling fetching graph search validity cache for NFT Groups
which was disabled previously.