check if the token graph should be reset
With graph search enabled, some jobs would finish and leave
some token graph nodes in an undetermined validity state.
This would cause subsequent validation jobs to fail.  In order
to prevent this problem, we can check the token graph for
this situation at the beginning of the next validation job.