refactor nft1 validation
Now we are using a new go-like WaitGroup class to
manage the validation of NFT1 children which requires
validation of NFT parents before the NFT can be validated.
The NFT1 validator uses `self.wg.wait()` to allow NFT genesis
and group parent download threads to complete before
continuing with the NFT validation.